+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.1.2](https://github.com/geops/openlayers-editor/compare/v2.1.1...v2.1.2) (2022-08-11)
+
+
+### Bug Fixes
+
+* **DrawControl:** add drawInteractionOptions for DrawControl ([#239](https://github.com/geops/openlayers-editor/issues/239)) ([e1ceb7c](https://github.com/geops/openlayers-editor/commit/e1ceb7c0a62c36658d231236be84773890d2e609))
+
+### [2.1.1](https://github.com/geops/openlayers-editor/compare/v2.1.0...v2.1.1) (2022-08-10)
+
+
+### Bug Fixes
+
+* **CAD:** snap enabled on edit feature vertices, CAD support for rotated maps ([#234](https://github.com/geops/openlayers-editor/issues/234)) ([0d706af](https://github.com/geops/openlayers-editor/commit/0d706af65bce2759e61dbae5d9ca359940251573))
+
 ## [2.1.0](https://github.com/geops/openlayers-editor/compare/v2.0.1...v2.1.0) (2022-06-20)